+++ test/Settings_test.cpp
@@ -1,7 +1,7 @@
 /**
  * @file test/Settings_test.cpp
  *
- * @date 2015-12-07
+ * @date 2016-01-21
  *
  * @author Youri Hoogstrate
  *
@@ -68,12 +68,12 @@ BOOST_AUTO_TEST_SUITE(Testing)
  *
  * @test
  *
- * @date 2015-06-05
+ * @date 2016-01-21
  */
 BOOST_AUTO_TEST_CASE(Test1)
 {
 	char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "ACTGactgACUGacug", nullptr};
-	int argc = sizeof(argv) / sizeof(char *) - 1;
+	signed int argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 	
 	Sequence sequence;
 	
@@ -81,7 +81,7 @@ BOOST_AUTO_TEST_CASE(Test1)
 	
 	BOOST_REQUIRE_EQUAL(sequence.size() , 16);
 	
-	int i = 0;
+	size_t i = 0;
 	
 	BOOST_CHECK_EQUAL(sequence[i++], Nucleotide::A);
 	BOOST_CHECK_EQUAL(sequence[i++], Nucleotide::C);
@@ -111,7 +111,7 @@ BOOST_AUTO_TEST_CASE(Test1)
  *
  * @test
  *
- * @date 2015-06-22
+ * @date 2016-01-21
  *
  * @todo Implement the possibility to run multiple entries from a FASTA file
  */
@@ -126,7 +126,7 @@ BOOST_AUTO_TEST_CASE(Test2)
 	
 	
 	char *argv[] = { (char *) PACKAGE_NAME, (char *) "-f", (char *) filename.c_str(), nullptr};
-	int argc = sizeof(argv) / sizeof(char *) - 1;
+	signed int argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 	
 	Sequence sequence;
 	
@@ -134,7 +134,7 @@ BOOST_AUTO_TEST_CASE(Test2)
 	
 	BOOST_REQUIRE_EQUAL(sequence.size() , 16);
 	
-	int i = 0;
+	size_t i = 0;
 	
 	BOOST_CHECK_EQUAL(sequence[i++], Nucleotide::A);
 	BOOST_CHECK_EQUAL(sequence[i++], Nucleotide::C);
@@ -171,13 +171,13 @@ BOOST_AUTO_TEST_CASE(Test2)
 BOOST_AUTO_TEST_CASE(Test3)
 {
 	Sequence sequence;
-	int argc;
+	signed int argc;
 	std::string is;
 	
 	{
 		// Check default value
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -187,7 +187,7 @@ BOOST_AUTO_TEST_CASE(Test3)
 	{
 		// Check enabling
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-p", (char *) "1", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -197,7 +197,7 @@ BOOST_AUTO_TEST_CASE(Test3)
 	{
 		// Check disabling
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-p", (char *) "0", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -205,12 +205,14 @@ BOOST_AUTO_TEST_CASE(Test3)
 	}
 }
 
+
+
 /**
  * @brief Tests whether the minimal hairpin-size can be set (-h)
  *
  * @test
  *
- * @date 2015-06-05
+ * @date 2016-01-21
  */
 BOOST_AUTO_TEST_CASE(Test4)
 {
@@ -221,7 +223,7 @@ BOOST_AUTO_TEST_CASE(Test4)
 	
 	// Check default value
 	char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", nullptr};
-	argc = sizeof(argv) / sizeof(char *) - 1;
+	argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 	
 	Settings settings = Settings(argc, argv, sequence);
 	
@@ -229,17 +231,17 @@ BOOST_AUTO_TEST_CASE(Test4)
 	
 	
 	// Check argumented values
-	for(signed int i = 0; i < 100; i ++)
+	for(unsigned int i = 0; i < 100; i ++)
 	{
 		is = std::to_string(i);
 		char *ics = (char *) is.c_str();
 		
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-h", (char *) ics, nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
-		BOOST_CHECK_MESSAGE(settings.minimal_hairpin_length  == i, "Failed to obtain min_hairpin_size of " << i);
+		BOOST_CHECK_MESSAGE(settings.minimal_hairpin_length == i, "Failed to obtain min_hairpin_size of " << i);
 	}
 }
 
@@ -258,7 +260,7 @@ BOOST_AUTO_TEST_CASE(Test5)
 	{
 		// Check example file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-x", (char *) "share/segmentation-fold/" SEGMENTS_FILE, nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -268,7 +270,7 @@ BOOST_AUTO_TEST_CASE(Test5)
 	{
 		// Check non existing file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-x", (char *) "/dev/null/neverexist", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		BOOST_CHECK_THROW(Settings settings = Settings(argc, argv, sequence), std::invalid_argument);
 	}
@@ -289,7 +291,7 @@ BOOST_AUTO_TEST_CASE(Test6)
 	{
 		// Check example file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -299,7 +301,7 @@ BOOST_AUTO_TEST_CASE(Test6)
 	{
 		// Check non existing file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-V", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -322,7 +324,7 @@ BOOST_AUTO_TEST_CASE(Test7)
 	{
 		// Check example file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -332,7 +334,7 @@ BOOST_AUTO_TEST_CASE(Test7)
 	{
 		// Check non existing file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "--version", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -355,7 +357,7 @@ BOOST_AUTO_TEST_CASE(Test8)
 	{
 		// Check example file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -365,7 +367,7 @@ BOOST_AUTO_TEST_CASE(Test8)
 	{
 		// Check non existing file
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "--help", nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
 		
@@ -378,7 +380,7 @@ BOOST_AUTO_TEST_CASE(Test8)
  *
  * @test
  *
- * @date 2015-07-20
+ * @date 2016-01-21
  */
 BOOST_AUTO_TEST_CASE(Test9)
 {
@@ -389,7 +391,7 @@ BOOST_AUTO_TEST_CASE(Test9)
 	
 	// Check default value
 	char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", nullptr};
-	argc = sizeof(argv) / sizeof(char *) - 1;
+	argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 	
 	Settings settings = Settings(argc, argv, sequence);
 	
@@ -397,13 +399,13 @@ BOOST_AUTO_TEST_CASE(Test9)
 	
 	
 	// Check argumented values
-	for(signed int i = 0; i < 100; i ++)
+	for(unsigned int i = 0; i < 100; i ++)
 	{
 		is = std::to_string(i);
 		char *ics = (char *) is.c_str();
 		
 		char *argv[] = { (char *) PACKAGE_NAME, (char *) "-s", (char *) "a", (char *) "-t", (char *) ics, nullptr};
-		argc = sizeof(argv) / sizeof(char *) - 1;
+		argc = (signed int) sizeof(argv) / (signed int) sizeof(char *) - 1;
 		
 		Settings settings = Settings(argc, argv, sequence);
